+++ packages/shared/src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './date.js';
-export * from './currency.js';
 export * from './pagination.js';
 export * from './response.js';